library/requirements.txt
amqp==5.3.1
asgiref==3.8.1
bcp47==0.1.0
billiard==4.2.1
celery==5.5.3
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
dj-rest-auth==7.0.1
Django==5.2.3
django-allauth==65.9.0
django-celery-beat==2.8.1
django-cors-headers==4.7.0
django-filter==25.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.10
gevent==25.5.1
greenlet==3.2.3
idna==3.10
inflection==0.5.1
iso-639==0.4.5
kombu==5.5.4
numpy==2.3.0
packaging==25.0
pandas==2.3.0
prompt_toolkit==3.0.51
pycparser==2.22
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
redis==6.2.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
zope.event==5.0
zope.interface==7.2
